DN-4: Individual comment component.
